reddit-analyzer/requirements.txt
# requirements.txt
flask==2.3.2
werkzeug==2.3.6
praw==7.7.0
pandas==1.5.3
pymongo
python-dotenv